fix(test): fix typing issues (#587)

+++ tests/functional/nccl_message_transfer.c
@@ -51,12 +51,12 @@ int main(int argc, char* argv[])
 	   and smaller than recv size. And check values 1. below
 	   the eager threshold, 2. between eager and rr threshold,
 	   and 3. above the rr threshold. */
-	size_t send_sizes[] = {4 * 1024, 16 * 1024, 1024 * 1024,
-	                       5 * 1024, 17 * 1024, 2 * 1024 * 1024,
-	                       4 * 1024, 16 * 1024, 1024 * 1024};
-	size_t recv_sizes[] = {4 * 1024, 16 * 1024, 1024 * 1024,
-	                       4 * 1024, 16 * 1024, 1024 * 1024,
-	                       5 * 1024, 17 * 1024, 2 * 1024 * 1024};
+	int send_sizes[] = {4 * 1024, 16 * 1024, 1024 * 1024,
+	                    5 * 1024, 17 * 1024, 2 * 1024 * 1024,
+	                    4 * 1024, 16 * 1024, 1024 * 1024};
+	int recv_sizes[] = {4 * 1024, 16 * 1024, 1024 * 1024,
+	                    4 * 1024, 16 * 1024, 1024 * 1024,
+	                    5 * 1024, 17 * 1024, 2 * 1024 * 1024};
 
 	/* For grouped recvs */
 	int tag = 1;
